removed automatic format determination
