+++ src/constants.ts
@@ -1,3 +1,7 @@
-import { Format } from './types';
+export enum Format {
+  TIMESERIES = 0,
+  TABLE = 1,
+  LOGS = 2,
+}
 
 export const DEFAULT_QUERY_FORMAT = Format.TABLE;
